Remove the "skip init schema" stuff
When SKIP_INIT_SCHEMA was unset, we'd spin up a postgres on the new
data directory, but it wasn't actually used.  Drop this, as the
loading of the initial schema is already handled by the prepare.sql in
the postgres --single call preceding this block.  This is likely the
reason sometimes Postgres doesn't want to shut down cleanly, as the
kill command in this block might trigger an unclean shutdown.